% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lebedev.r
\name{cub_lebedev}
\alias{cub_lebedev}
\title{quadrature_sphere}
\usage{
cub_lebedev(Nq = 30)
}
\arguments{
\item{Nq}{number of integration points}

\item{quadrature}{quadrature method, using either Gauss Legendre quadrature (default), Quasi Monte Carlo, regular grid, or "cheap" (3 axes)}

\item{init}{(qmc method only) logical: restart, or continue from previous call}
}
\description{
Quadrature points on a sphere
}
\details{
Numerical integration points for angular averaging
}
\seealso{
Other low_level quadrature: \code{\link{quadrature_sphere}}
}
\author{
baptiste Auguie
}
\concept{low_level quadrature}
