% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{lebedev_table}
\alias{lebedev_table}
\alias{sphericaldesigns_table}
\alias{gl_table}
\alias{grid_table}
\alias{qmc_table}
\alias{random_table}
\alias{fibonacci_table}
\title{lebedev_table}
\format{
data.frame with 2 columns
\describe{
  \item{degree}{}
  \item{N}{}
}

data.frame with 2 columns
\describe{
  \item{degree}{}
  \item{N}{}
}

vector

vector

vector

vector

vector
}
\source{
\url{https://people.sc.fsu.edu/~jburkardt/c_src/sphere_lebedev_rule/sphere_lebedev_rule.html}

\url{https://people.sc.fsu.edu/~jburkardt/c_src/sphere_lebedev_rule/sphere_lebedev_rule.html}
}
\usage{
lebedev_table

sphericaldesigns_table

gl_table

grid_table

qmc_table

random_table

fibonacci_table
}
\description{
nodes and weights

nodes and weights

valid numbers

valid numbers

valid numbers

valid numbers

valid numbers
}
\keyword{datasets}
