% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubs-package.r
\docType{package}
\name{cubs-package}
\alias{cubs-package}
\title{cubs}
\description{
Cubature methods for orientation averaging.
}
\details{
Cubature methods for orientation averaging.
}
\author{
baptiste Auguie
}
\keyword{package}
\keyword{packagelibrary}
