% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubature.r
\name{cubs}
\alias{cubs}
\title{cubature_sphere}
\usage{
cubs(
  N = 30,
  cubature = c("lebedev", "sphericaldesigns", "gl", "fibonacci", "grid", "qmc",
    "random")
)
}
\arguments{
\item{N}{number of integration points}

\item{cubature}{cubature method, using either Gauss Legendre cubature (default), Quasi Monte Carlo, regular grid, or "cheap" (3 axes)}
}
\description{
cubature points on a sphere
}
\details{
Numerical integration points for angular averaging
}
\author{
baptiste Auguie
}
\concept{low_level cubature}
