% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{sphericaldesigns}
\alias{sphericaldesigns}
\title{spherical designs}
\format{
data.frame with 3 columns
\describe{
  \item{phi}{azmuth}
  \item{theta}{polar angle}
  \item{weight}{weight}
}
}
\source{
\url{https://web.maths.unsw.edu.au/~rsw/Sphere/EffSphDes/}
}
\usage{
sphericaldesigns
}
\description{
Nodes and weights of spherical t-designs
}
\keyword{datasets}
