% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_GL}
\alias{is_valid_GL}
\alias{is_valid_grid}
\alias{is_valid_fibonacci}
\title{is_valid_GL}
\usage{
is_valid_GL(N)

is_valid_grid(N)

is_valid_fibonacci(N)
}
\arguments{
\item{N}{number of integration points}
}
\description{
cubature points on a sphere
}
\details{
Tests if a number of points is valid
}
\author{
baptiste Auguie
}
\concept{low_level}
