% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{lebedev}
\alias{lebedev}
\title{lebedev}
\format{
data.frame with 3 columns
\describe{
  \item{phi}{}
  \item{theta}{}
  \item{weight}{}
}
}
\source{
\url{https://people.sc.fsu.edu/~jburkardt/c_src/sphere_lebedev_rule/sphere_lebedev_rule.html}
}
\usage{
lebedev
}
\description{
nodes and weights
}
\keyword{datasets}
