% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{sphericaldesigns}
\alias{sphericaldesigns}
\title{spherical designs}
\format{
data.frame with 3 columns
\describe{
  \item{phi}{}
  \item{theta}{}
  \item{weight}{}
}
}
\source{
\url{}
}
\usage{
sphericaldesigns
}
\description{
nodes and weights
}
\keyword{datasets}
